requirements.txt
streamlit>=1.24.0
python-docx
openai
python-dotenv
pandas

